feat(Auto-Align): Added the auto align face offset to the robot
The face that is wanted to be facing the reef may not always be 0_deg
from the robot. This commit accounts for any offset that the face needs
to be at. So if we want to align the side of the robot to the reef
instead of the front of the robot you add an angle offset such that the
face you want faces the reef.