feat(Auto-Align): Added initial velocity for PID Auto Align Controllers
Instead of just resetting the PID controllers to only have a position,
either distance for y_controller or radians for rot_controller, we also
include a velocity into that reset function. This should allow us to be
able to track with relative position in any state even when we were
initially moving.